+++ src/test/java/simple/TestSimple.java
@@ -18,21 +18,21 @@ public class TestSimple {
 		p.Parse(INTEGER, 10);
 		p.Parse(TIMES, 0);
 		p.Parse(INTEGER, 4);
-		p.Parse(0, 0);
+		p.Parse((short)0, 0);
 		assertEquals("(5 PLUS (10 TIMES 4))", r.expr.String());
 		r.expr = null;
 
 		p.Parse(INTEGER, 15);
 		p.Parse(DIVIDE, 0);
 		p.Parse(INTEGER, 5);
-		p.Parse(0, 0);
+		p.Parse((short)0, 0);
 		assertEquals("(15 DIVIDE 5)", r.expr.String());
 		r.expr = null;
 
 		p.Parse(INTEGER, 50);
 		p.Parse(PLUS, 0);
 		p.Parse(INTEGER, 125);
-		p.Parse(0, 0);
+		p.Parse((short)0, 0);
 		assertEquals("(50 PLUS 125)", r.expr.String());
 		r.expr = null;
 
@@ -41,7 +41,7 @@ public class TestSimple {
 		p.Parse(INTEGER, 125);
 		p.Parse(PLUS, 0);
 		p.Parse(INTEGER, 125);
-		p.Parse(0, 0);
+		p.Parse((short)0, 0);
 		assertEquals("((50 TIMES 125) PLUS 125)", r.expr.String());
 		r.expr = null;
 	}
